requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apturl==0.5.2
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brlapi==0.8.5
Brotli==1.1.0
bs4==0.0.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.8
contourpy==1.3.1
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.3.2
defer==1.0.6
deprecation==2.0.7
distro==1.9.0
emoji==2.14.0
eval_type_backport==0.2.0
eyed3==0.9.7
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
groq==0.13.0
gyp==0.1
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.6
ifaddr==0.2.0
IMDbPY==2021.4.18
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kaggle==1.6.17
kiwisolver==1.4.7
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-ollama==0.2.1
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.147
launchpadlib==1.11.0
layoutparser==0.3.4
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mutagen==1.46.0
mypy==1.13.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
ollama==0.4.4
onboard==1.4.1
onnx==1.17.0
onnxruntime==1.20.1
openai==1.57.1
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.0
PAM==0.4.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.4
pexpect==4.9.0
pillow==10.2.0
pillow_heif==0.21.0
pinecone==5.4.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
portalocker==3.0.0
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.8
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.25.1
pycryptodome==3.21.0
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyparted==3.12.0
pypdf==5.1.0
pypdfium2==4.30.0
pypng==0.20231004.0
python-apt==2.7.7+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-gnupg==0.5.2
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
python-slugify==8.0.4
python-xlib==0.33
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
qrcode==7.4.2
RapidFuzz==3.10.1
regex==2024.11.6
repolib==2.2.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setproctitle==1.3.3
setuptools==68.1.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
sympy==1.13.1
systemd-python==235
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.12
tinycss2==1.2.1
tldextract==3.1.2
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ufw==0.36.2
Unidecode==1.3.8
unstructured==0.16.10
unstructured-client==0.28.1
unstructured-inference==0.8.1
urllib3==2.0.7
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.42.0
wrapt==1.17.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
yarl==1.18.3
yt-dlp==2024.4.9
